+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.3.0
+- Cron Health / Missed Schedules Site Health check
+- Updated documentation to reflect new check
+
 ### 0.2.0
 - Added WordPress core version diagnostic check
 - Updated documentation to reflect new check